requirements.txt
imageio==2.4.1
matplotlib==2.1.1
numpy==1.16.2
pyserial==3.4
scikit-image==0.14.2
scipy==1.2.1
six==1.11.0
Adafruit-BluefruitLE==0.9.10
dash==0.42.0
plotly==3.10
future==0.18.2
PyObjC==5.1; sys_platform=="darwin"
